modified error message
