+++ src/controllers/user.js
@@ -1,8 +1,9 @@
+/* eslint-disable max-classes-per-file */
 import bcrypt from "bcrypt";
 import Util from "../utilities/util";
 import User from "../services/UserService/User";
 import jwtHelper from "../utilities/Jwt";
-import { registerValidation } from "../validation/userValidation";
+import { registerValidation, loginValidation } from "../validation/userValidation";
 
 const { generateToken } = jwtHelper;
 const util = new Util();
@@ -34,5 +35,31 @@ export default class userController {
       util.setError(400, error.message);
       throw util.send(res);
     }
+  }
+
+  static async loginUser(req, res) {
+    try {
+      const { error } = loginValidation(req.body);
+      if (error) {
+        util.setError(400, "Validation Error", error.message);
+        return util.send(res);
+      }
+      const { email, username, password } = req.body;
+      const user = await User.emailExist(email);
+      if (username) {
+        return res.status(400).send({ message: "Username is not required" });
+      }
+      if (!user) {
+        return res.status(400).send({ message: "Email does not exist." });
+      }
+      const validpass = await bcrypt.compare(password, user.password);
+      if (!validpass) {
+        return res.status(400).send({ message: "Password is not correct!." });
+      }
+      const token = await generateToken({ user });
+      util.setSuccess(201, "User Logged in!", token);
+    } catch (error) {
+      util.setError(400, error.message);
+    } return util.send(res);
   }
 }
