+++ src/routes/userRoutes.js
@@ -1,9 +1,8 @@
 import { Router } from "express";
 import userController from "../controllers/user";
-import loginController from "../controllers/user-sign-in";
 
 const router = Router();
 router.post("/users/signup", userController.createUser);
-router.post("/users/signin", loginController.loginUser);
+router.post("/users/signin", userController.loginUser);
 
 export default router;
