+++ src/validation/userValidation.js
@@ -15,6 +15,7 @@ const registerValidation = user => {
       .empty()
       .messages({
         "string.email": "Please enter a valid email",
+        "string.empty": "Sorry, Email cannot be an empty field",
       }),
     password: Joi.string().required().empty().min(5)
       .max(1024)
@@ -30,14 +31,23 @@ const registerValidation = user => {
 
 const loginValidation = user => {
   const schema = Joi.object({
+    username: Joi.string().alphanum().min(3)
+      .max(255)
+      .empty()
+      .messages({
+        "string.alphanum": "Sorry, Username is not required",
+        "string.empty": "Sorry, username is not required",
+        "string.min": "Sorry, username is not required"
+      }),
     email: Joi.string().required().email({ minDomainSegments: 2, tlds: { allow: ["com", "net", "uk", "co", "io"] } }).min(5)
       .max(100)
       .empty()
       .messages({
         "string.email": "Please enter a valid email",
+        "string.empty": "Sorry, Email cannot be an empty field",
       }),
     password: Joi.string().required().min(5).max(1024)
-    .pattern(new RegExp("^[a-zA-Z0-9]{3,30}$"))
+      .pattern(new RegExp("^[a-zA-Z0-9]{3,30}$"))
       .messages({
         "string.pattern.base": "Password must contain only alphanumeric characters.",
         "string.empty": "Sorry, password cannot be an empty field",
